Only create a tmp dir if there's a new deployment.
Nasty bug this one. The script was creating a temporary directory in
/tmp on each run, instead of just when needed (ie; if there's a new
deployment).

This means that if you don't have a cron to purge /tmp, you can reach
the maximum number of allowed items in that directory in just a few
weeks and put /tmp out of use. This can cause other things to break.

The fix is to just ensure we only create the tmp directory when it's
needed.